+++ packages/react-ui-validations/rollup.react-ui.config.js
@@ -1,2 +1,2 @@
-import { buildConfig } from "./rollup.base.config";
-export default buildConfig("react-ui-dist", "@skbkontur/react-ui");
+import { buildConfig } from './rollup.base.config';
+export default buildConfig('react-ui-dist', '@skbkontur/react-ui');