+++ packages/retail-ui/components/Popup/Popup.tsx
@@ -3,7 +3,6 @@ import * as React from 'react';
 import { findDOMNode } from 'react-dom';
 import * as PropTypes from 'prop-types';
 import RenderContainer from '../RenderContainer';
-import RenderLayer from '../RenderLayer';
 import ZIndex from '../ZIndex';
 import { Transition } from 'react-transition-group';
 import raf from 'raf';
@@ -16,6 +15,7 @@ import { Nullable } from '../../typings/utility-types';
 import warning from 'warning';
 import { FocusEventType, MouseEventType } from '../../typings/event-types';
 import { isFunction } from '../../lib/utils';
+import LifeCycleProxy from '../internal/LifeCycleProxy';
 
 const POPUP_BORDER_DEFAULT_COLOR = 'transparent';
 const TRANSITION_TIMEOUT = { enter: 0, exit: 200 };
@@ -80,7 +80,6 @@ export interface PopupProps extends PopupHandlerProps {
   pinSize: number;
   popupOffset: number;
   positions: PopupPosition[];
-  onCloseRequest?: () => void;
   useWrapper: boolean;
   ignoreHover?: boolean;
 }
@@ -109,7 +108,7 @@ export default class Popup extends React.Component<PopupProps, PopupState> {
      */
     backgroundColor: PropTypes.string,
 
-    children: PropTypes.oneOfType([PropTypes.node, PropTypes.func]).isRequired,
+    children: PropTypes.oneOfType([PropTypes.node, PropTypes.func]),
 
     /**
      * Показывать ли пин
@@ -172,9 +171,7 @@ export default class Popup extends React.Component<PopupProps, PopupState> {
     useWrapper: false,
   };
 
-  public state: PopupState = {
-    location: null,
-  };
+  public state: PopupState = { location: null };
 
   private layoutEventsToken: Nullable<ReturnType<typeof LayoutEvents.addListener>>;
   private locationUpdateId: Nullable<number> = null;
@@ -187,7 +184,7 @@ export default class Popup extends React.Component<PopupProps, PopupState> {
     this.layoutEventsToken = LayoutEvents.addListener(this.handleLayoutEvent);
   }
 
-  public componentWillReceiveProps(nextProps: PopupProps) {
+  public componentWillReceiveProps(nextProps: Readonly<PopupProps>) {
     const isGoingToOpen = !this.props.opened && nextProps.opened;
     const isGoingToUpdate = this.props.opened && nextProps.opened;
     const isGoingToClose = this.props.opened && !nextProps.opened;
@@ -206,14 +203,6 @@ export default class Popup extends React.Component<PopupProps, PopupState> {
     }
   }
 
-  public componentDidUpdate(prevProps: PopupProps, prevState: PopupState) {
-    const hadNoLocation = prevState.location === null;
-    const hasLocation = this.state.location !== null;
-    if (hadNoLocation && hasLocation && this.props.onOpen) {
-      this.props.onOpen();
-    }
-  }
-
   public componentWillUnmount() {
     this.cancelDelayedUpdateLocation();
     this.removeEventListeners(this.anchorElement);
@@ -224,8 +213,7 @@ export default class Popup extends React.Component<PopupProps, PopupState> {
   }
 
   public render() {
-    const { useWrapper, onCloseRequest, opened } = this.props;
-    const anchorElement = this.props.anchorElement;
+    const { anchorElement, useWrapper } = this.props;
 
     let child: Nullable<React.ReactNode> = null;
     if (anchorElement instanceof HTMLElement) {
@@ -237,19 +225,9 @@ export default class Popup extends React.Component<PopupProps, PopupState> {
     }
 
     return (
-      <RenderLayer
-        onClickOutside={this.handleClickOutside}
-        onFocusOutside={this.handleFocusOutside}
-        /**
-         * If onCloseRequest is not specified handleClickOutside and handleFocusOutside
-         * are doing nothing. So there is no need in RenderLayer at all.
-         */
-        active={Boolean(onCloseRequest) && opened}
-      >
-        <RenderContainer anchor={child} ref={child ? this.refAnchorElement : undefined}>
-          {this.renderContent()}
-        </RenderContainer>
-      </RenderLayer>
+      <RenderContainer anchor={child} ref={child ? this.refAnchorElement : undefined}>
+        {this.renderContent()}
+      </RenderContainer>
     );
   }
 
@@ -325,52 +303,63 @@ export default class Popup extends React.Component<PopupProps, PopupState> {
   };
 
   private renderContent() {
+    const props = this.props;
+    const children = this.renderChildren();
+
+    if (!props.opened || !children) {
+      return null;
+    }
+
     const location = this.state.location || DUMMY_LOCATION;
     const { direction } = PopupHelper.getPositionObject(location.position);
-    const { backgroundColor } = this.props;
+    const { backgroundColor, disableAnimations } = props;
     const rootStyle: React.CSSProperties = {
       top: location.coordinates.top,
       left: location.coordinates.left,
-      maxWidth: this.props.maxWidth,
+      maxWidth: props.maxWidth,
     };
 
+    // This need to correct handle order of lifecycle hooks with portal and react@15
+    // For more details see issue #1257
     return (
-      <Transition
-        timeout={TRANSITION_TIMEOUT}
-        appear={!this.props.disableAnimations}
-        in={this.props.opened}
-        mountOnEnter
-        unmountOnExit
-        enter={!this.props.disableAnimations}
-        exit={!this.props.disableAnimations}
-      >
-        {(state: string) => (
-          <ZIndex
-            key={this.state.location ? 'real' : 'dummy'}
-            delta={1000}
-            ref={this.refPopupElement}
-            className={cn({
-              [styles.popup]: true,
-              [styles['popup-ignore-hover']]: this.props.ignoreHover,
-              [styles.shadow]: this.props.hasShadow,
-              [styles['transition-enter']]: state === 'entering',
-              [styles['transition-enter-active']]: state === 'entered',
-              [styles['transition-exit']]: state === 'exiting',
-              [styles[('transition-enter-' + direction) as keyof typeof styles]]: true,
-            })}
-            style={rootStyle}
-            onMouseEnter={this.handleMouseEnter}
-            onMouseLeave={this.handleMouseLeave}
-          >
-            <div className={styles.content}>
-              <div className={styles.contentInner} style={{ backgroundColor }}>
-                {this.renderChildren()}
+      <LifeCycleProxy onDidUpdate={this.handleDidUpdate} props={this.state}>
+        <Transition
+          timeout={TRANSITION_TIMEOUT}
+          appear={!disableAnimations}
+          in
+          mountOnEnter
+          unmountOnExit
+          enter={!disableAnimations}
+          exit={!disableAnimations}
+        >
+          {(state: string) => (
+            <ZIndex
+              key={this.state.location ? 'real' : 'dummy'}
+              delta={1000}
+              ref={this.refPopupElement}
+              className={cn({
+                [styles.popup]: true,
+                [styles['popup-ignore-hover']]: props.ignoreHover,
+                [styles.shadow]: props.hasShadow,
+                [styles['transition-enter']]: state === 'entering',
+                [styles['transition-enter-active']]: state === 'entered',
+                [styles['transition-exit']]: state === 'exiting',
+                [styles[('transition-enter-' + direction) as keyof typeof styles]]: true,
+              })}
+              style={rootStyle}
+              onMouseEnter={this.handleMouseEnter}
+              onMouseLeave={this.handleMouseLeave}
+            >
+              <div className={styles.content}>
+                <div className={styles.contentInner} style={{ backgroundColor }}>
+                  {children}
+                </div>
               </div>
-            </div>
-            {this.renderPin(location.position)}
-          </ZIndex>
-        )}
-      </Transition>
+              {this.renderPin(location.position)}
+            </ZIndex>
+          )}
+        </Transition>
+      </LifeCycleProxy>
     );
   }
 
@@ -410,14 +399,6 @@ export default class Popup extends React.Component<PopupProps, PopupState> {
     );
   }
 
-  private handleClickOutside = () => {
-    this.requestClose();
-  };
-
-  private handleFocusOutside = () => {
-    this.requestClose();
-  };
-
   private handleLayoutEvent = () => {
     if (this.anchorInstance) {
       this.updateAnchorElement(this.extractElement(this.anchorInstance));
@@ -427,9 +408,17 @@ export default class Popup extends React.Component<PopupProps, PopupState> {
     }
   };
 
+  private handleDidUpdate = (prevProps: PopupState, props: PopupState) => {
+    const hadNoLocation = prevProps.location === null;
+    const hasLocation = props.location !== null;
+    if (hadNoLocation && hasLocation && this.props.onOpen) {
+      this.props.onOpen();
+    }
+  };
+
   private delayUpdateLocation() {
     this.cancelDelayedUpdateLocation();
-    this.locationUpdateId = raf(() => this.updateLocation());
+    this.locationUpdateId = raf(this.updateLocation);
   }
 
   private cancelDelayedUpdateLocation() {
@@ -439,7 +428,7 @@ export default class Popup extends React.Component<PopupProps, PopupState> {
     }
   }
 
-  private updateLocation() {
+  private updateLocation = () => {
     const popupElement = this.lastPopupElement;
 
     if (!popupElement) {
@@ -450,17 +439,11 @@ export default class Popup extends React.Component<PopupProps, PopupState> {
     if (!this.locationEquals(this.state.location, location)) {
       this.setState({ location });
     }
-  }
+  };
 
   private resetLocation = () => {
     this.cancelDelayedUpdateLocation();
     this.setState({ location: null });
-  };
-
-  private requestClose = () => {
-    if (this.props.onCloseRequest) {
-      this.props.onCloseRequest();
-    }
   };
 
   private locationEquals(x: Nullable<PopupLocation>, y: Nullable<PopupLocation>) {