+++ packages/retail-ui/components/Tooltip/Tooltip.tsx
@@ -154,9 +154,7 @@ class Tooltip extends React.Component<TooltipProps, TooltipState> {
 
   private static triggersWithoutCloseButton: TooltipTrigger[] = ['hover', 'hoverAnchor', 'focus'];
 
-  public state = {
-    opened: false,
-  };
+  public state: TooltipState = { opened: false };
 
   private hoverTimeout: Nullable<number> = null;
   private contentElement: Nullable<HTMLElement> = null;
@@ -181,23 +179,13 @@ class Tooltip extends React.Component<TooltipProps, TooltipState> {
   }
 
   public render() {
-    const { popupProps, layerProps } = this.getProps();
     const props = this.props;
-    const children = props.children;
     const content = this.renderContent();
-    const anchorElement = children || props.anchorElement;
+    const { popupProps, layerProps = { active: false } } = this.getProps();
+    const anchorElement = props.children || props.anchorElement;
+    const popup = this.renderPopup(anchorElement, popupProps, content);
 
-    if (content != null) {
-      const popup = this.renderPopup(anchorElement, popupProps, content);
-
-      return <RenderLayer {...layerProps}>{popup}</RenderLayer>;
-    }
-
-    if (React.isValidElement(children)) {
-      return children;
-    }
-
-    return <span>{children}</span>;
+    return <RenderLayer {...layerProps}>{popup}</RenderLayer>;
   }
 
   public renderContent = () => {
@@ -274,7 +262,7 @@ class Tooltip extends React.Component<TooltipProps, TooltipState> {
   }
 
   private getProps(): {
-    layerProps: Partial<RenderLayerProps>;
+    layerProps?: Partial<RenderLayerProps>;
     popupProps: Partial<PopupProps>;
   } {
     const props = this.props;
@@ -294,9 +282,6 @@ class Tooltip extends React.Component<TooltipProps, TooltipState> {
 
       case 'closed':
         return {
-          layerProps: {
-            active: false,
-          },
           popupProps: {
             opened: false,
             useWrapper,
@@ -306,9 +291,6 @@ class Tooltip extends React.Component<TooltipProps, TooltipState> {
       case 'hoverAnchor':
       case 'hover':
         return {
-          layerProps: {
-            active: false,
-          },
           popupProps: {
             onMouseEnter: this.handleMouseEnter,
             onMouseLeave: this.handleMouseLeave,
@@ -330,9 +312,6 @@ class Tooltip extends React.Component<TooltipProps, TooltipState> {
 
       case 'focus':
         return {
-          layerProps: {
-            active: false,
-          },
           popupProps: {
             onFocus: this.handleFocus,
             onBlur: this.handleBlur,